+++ src/org/tizzer/smmgr/system/model/response/QuerySomeStoreResponseDto.java
@@ -41,12 +41,12 @@ public class QuerySomeStoreResponseDto extends ResultResponse {
 
     static class DataSet {
 
-        private Long id;
+        private Integer id;
         private String name;
         private String address;
         private Date foundDate;
 
-        public void setId(Long id) {
+        public void setId(Integer id) {
             this.id = id;
         }
